requirements.txt
aiohttp==3.7.4.post0
aniso8601==7.0.0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.5
backcall==0.2.0
behave==1.2.6
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cucumber-tag-expressions==3.0.1
decorator==4.4.2
Faker==8.1.4
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
lxml==4.6.3
multidict==5.1.0
mypy-extensions==0.4.3
parse==1.19.0
parse-type==0.5.2
parso==0.8.1
pathspec==0.8.1
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.16
pycodestyle==2.6.0
pyderman==2.0.1
Pygments==2.7.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-graphql-client==0.4.3
pytils==0.3
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
Rx==1.6.1
selenium==3.141.0
six==1.15.0
text-unidecode==1.3
toml==0.10.2
tools==0.1.9
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webdriver-manager==3.3.0
websockets==9.0.2
yarl==1.6.3
